backend/requirements.txt
aiosmtpd==1.4.6
alembic==1.16.4
amqp==5.3.1
aniso8601==10.0.1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
atpublic==6.0.1
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
cssselect2==0.8.0
Flask==3.1.1
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
oscrypto==1.3.0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
pycparser==2.22
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
pypdf==5.8.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
reportlab==4.4.3
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.41
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.17
